pyproject.toml
[tool.poetry]
name = "gpt-your-data"
version = "0.1.0"
description = ""
authors = ["Henrique Breim <henrique@breim.com.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
faiss-cpu = "^1.8.0.post1"
transformers = "^4.44.0"
torch = "^2.4.0"
fastapi = "^0.112.1"
uvicorn = "^0.30.6"
sqlalchemy = "^2.0.32"
beautifulsoup4 = "^4.12.3"
requests = "^2.32.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "gpt_your_data.main:start"
crawl-data = "gpt_your_data.scripts.crawl_data:main"
create-tables = "gpt_your_data.config.migrations.create_episodes_migration:create_tables"
